fix JS event loop blocking
